refactored subsection important to take in children for dynamic rendering for our work page
