+++ pages/ourwork.js
@@ -1,6 +1,7 @@
 import Nav from '../components/nav'
 import Footer from '../components/footer'
 import Head from '../components/head'
+import SubsectionImportant from '../components/subsectionImportant'
 import '../static/styles/main.scss'
 
 const Ourwork = () => (
@@ -9,8 +10,11 @@ const Ourwork = () => (
     <Head title='Our Work' />
 
     <div className='body'>
-      <h1>Our Work</h1>
       <div className='description' />
+      <SubsectionImportant 
+        titleText="How we work"
+        descriptionText="We believe public school teachers don't receive the recognition they deserve. We're working to change that."
+      />
     </div>
     <Footer />
   </div>
